refined the tests for borrower.
